Additional refactoring and testing of current code with real ESP8266 device

+++ src/main/java/coolbeans/microthings8266hub/esp8266/Esp8266Commands.java
@@ -7,7 +7,7 @@ public enum Esp8266Commands {
 
     Esp8266Commands(int value) { this.value = value; }
 
-    public int getValue() {
-        return value;
+    public byte getValue() {
+        return (byte) value;
     }
 }
